+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## [v0.5.0]
+> December 10, 2018
+
+- Add support to mongoose 5.x (@joelmukuthu, @gforge)
+- Add `deleteById` static method #16
+- Add `countDocuments` method with related override methods (only for v5 Mongoose) #45
+
 ## [v0.4.0]
 > July 10, 2016